+++ src/coinbase/faucet_transaction.ts
@@ -1,6 +1,9 @@
 import { FaucetTransaction as FaucetTransactionModel } from "../client";
 import { TransactionStatus } from "./types";
+import { Coinbase } from "./coinbase";
 import { Transaction } from "./transaction";
+import { delay } from "./utils";
+import { TimeoutError } from "./errors";
 
 /**
  * Represents a transaction from a faucet.
@@ -78,6 +81,62 @@ export class FaucetTransaction {
    */
   public getAddressId(): string {
     return this.transaction.toAddressId()!;
+  }
+
+  /**
+   * Waits for the FaucetTransaction to be confirmed on the Network or fail on chain.
+   * Waits until the FaucetTransaction is completed or failed on-chain by polling at the given interval.
+   * Raises an error if the FaucetTransaction takes longer than the given timeout.
+   *
+   * @param options - The options to configure the wait function.
+   * @param options.intervalSeconds - The interval to check the status of the FaucetTransaction.
+   * @param options.timeoutSeconds - The maximum time to wait for the FaucetTransaction to be confirmed.
+   *
+   * @returns The FaucetTransaction object in a terminal state.
+   * @throws {Error} if the FaucetTransaction times out.
+   */
+  public async wait({
+    intervalSeconds = 0.2,
+    timeoutSeconds = 10,
+  } = {}): Promise<FaucetTransaction> {
+    const startTime = Date.now();
+
+    while (Date.now() - startTime < timeoutSeconds * 1000) {
+      await this.reload();
+
+      // If the FaucetTransaction is in a terminal state, return the FaucetTransaction.
+      if (this.transaction.isTerminalState()) {
+        return this;
+      }
+
+      await delay(intervalSeconds);
+    }
+
+    throw new TimeoutError("FaucetTransaction timed out");
+  }
+
+  /**
+   * Reloads the FaucetTransaction model with the latest data from the server.
+   *
+   * @returns {FaucetTransaction} The reloaded FaucetTransaction object.
+   * @throws {APIError} if the API request to get a FaucetTransaction fails.
+   */
+  public async reload(): Promise<FaucetTransaction> {
+    const result = await Coinbase.apiClients.externalAddress!.getFaucetTransaction(
+      this.transaction.getNetworkId(),
+      this.getAddressId(),
+      this.getTransactionHash(),
+    );
+
+    this.model = result?.data;
+
+    if (!this.model?.transaction) {
+      throw new Error("FaucetTransaction model cannot be empty");
+    }
+
+    this._transaction = new Transaction(this.model.transaction!);
+
+    return this;
   }
 
   /**