+++ src/tests/external_address_test.ts
@@ -603,6 +603,7 @@ describe("ExternalAddress", () => {
         address.getNetworkId(),
         address.getId(),
         undefined,
+        true, // Skip wait should be true.
       );
     });